refactor(GUI): integrate `etcher-latest-version` into the main repo
`etcher-latest-version` was kept in a separate repository in order to
re-use it with the Etcher website, however the Etcher website is not
using it at all, and we're moving towards having the website in the main
repository.

Therefore, this commit brings back the logic from
`etcher-latest-version`, but introduces it as `lib/shared/s3.js`, in
order to not tie ourselves to the AngularJS framework, and as a step
towards the Etcher SDK.

As a nice little bonus, this commit adds support for an
`ETCHER_LATEST_VERSION` environment variable that can be used to trick
Etcher that there is an available update, and therefore show the update
notifier modal.

See: https://github.com/resin-io-modules/etcher-latest-version
Fixes: https://github.com/resin-io/etcher/issues/953
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>